Shield Material Created
The deflection has been balanced to a state that is ready for testing.
Deflection collision now works as intended - so far no obvious bugs from
single player testing.
The material has been created, iterated upon and refined - just waiting
on polish on this first pass.
Updated the deflection in each of the Characters.
Created 2 dissolve materials that can be used later on breakable walls
but is also being used to experiment with the shield decay after it has
been used.
Shield now updates fringe colour to match that of the player's team
colour.