Door Hatch and Quick Kill SFX
Added WAVs and relevant cues for the sound of the hazard door opening
and closing, as well as the sound to be played when you get a succesful
quick kill and pickup in STS.

Max concurrency on door sounds has been set to 4. This may change
dependent on how many doors there are and how it sounds when testing.